Move Header make public install dart_api.h
