+++ CMakeLists.txt
@@ -22,6 +22,7 @@ set(OUTPUT_DEBUG "${OUTPUT_PATH_EXT}/Debug")
 set(OUTPUT_RELEASE "${OUTPUT_PATH_EXT}/Release")
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY "${PROJECT_SOURCE_DIR}/out/$<$<CONFIG:DEBUG>:${OUTPUT_DEBUG}>$<$<CONFIG:RELEASE>:${OUTPUT_RELEASE}>")
 set(EXECUTABLE_OUTPUT_PATH "${PROJECT_SOURCE_DIR}/out/${OUTPUT_PATH_EXT}/${CMAKE_BUILD_TYPE}")
+set(LIBRARY_OUTPUT_DIRECTORY "${PROJECT_SOURCE_DIR}/out/${OUTPUT_PATH_EXT}/${CMAKE_BUILD_TYPE}")
 
 
 add_subdirectory(src)
@@ -31,3 +32,4 @@ endif(BUILD_SMAPLES)
 
 MESSAGE(STATUS "Build to ${CMAKE_LIBRARY_OUTPUT_DIRECTORY}")
 MESSAGE(STATUS "Build to ${EXECUTABLE_OUTPUT_PATH}")
+MESSAGE(STATUS "Build to ${LIBRARY_OUTPUT_DIRECTORY}")
