+++ examples/simple_example/main.cpp
@@ -1,8 +1,7 @@
 #include <string.h>
 #include <iostream>
 
-#include "dart_dll.h"
-
+#include <dart_dll.h>
 #include <dart_api.h>
 
 Dart_Handle HandleError(Dart_Handle handle) {
